引入 Spring AI 并实现Agent核心交互层 集成查店与秒杀tool

+++ src/main/java/com/hmdp/service/impl/FollowServiceImpl.java
@@ -13,7 +13,7 @@ import com.hmdp.utils.UserHolder;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
@@ -69,7 +69,7 @@ public class FollowServiceImpl extends ServiceImpl<FollowMapper, Follow> impleme
         // 获取登录用户
         Long userId = UserHolder.getUser().getId();
         // 查询是否关注
-        Integer count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count();
+        Long count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count();
         return Result.ok(count > 0);
     }
